Ignore read-only/old jobs when checking conflicts
Currently, we disallow any enrollment-writing job
from starting if the associated program has any
jobs in the one of processing states (IN PROGRESS,
RETRYING, and PENDING). This commit relaxes that
constraint in two ways:
* Only block if a 'write' job is processing
* Ignore jobs of a certain age (currently 30 minutes)

The reason for the second relaxation is that jobs
occasionally becomes 'stuck' in the PENDING state, so
we have to modify the database to unblock future jobs
for that program. This is not desirable, so we can assume
that any job older than 30 minutes will never complete.

If Registrar begins to have longer-running jobs, then the
maximum age may need to be reconsidered.