Bulk-fetch programs from Discovery for performance
Currently, when the program details cache is cold,
we repeatedly call Course Discovery to fetch program
details.

This PR, instead, makes it so we fetch all non-cached
programs at once using the `uuids` filter on Discovery's
/api/v1/programs endpoint.

New functionality is behind feature flag, which is
enabled for devstack and tests but disabled (for now)
for stage and prod.

MST-191