Increase Discovery cache timeout to 24 hours
It was previously only 2 minutes due to belief
that the delay between modifying an external
course key in Publisher and seeing the result in
the Registrar API needed to be low.

However, the performance issues on Program Console
are currently a larger concern. We believe that
increasing this timeout will work as a stop-gap
to optimize the /api/v2/programs/ endpoint, which
Program Console relies upon for the initial page load.

MST-255