Update celery autodiscover mechanism for celery 4. (#381)
Pulling in INSTALLED_APPS and passing it explicitly was more necessary before Celery 4.  In celery 4 and later,
celery introduces an autodiscovery mechanism for django apps.  Update to the Celery 4 way because the other way
introduces a race condition in Celery 4 that can result in not all tasks getting properly discovered.

See https://github.com/edx/edx-platform/pull/25467 for an example of where this had to be fixed in edx-platform.