fix:add default transport so celery will run tasks eagerly
This makes sure we set a default transport, which is helpful when
we're running with CELERY_TASK_ALWAYS_EAGER. That setting enables us
to move forward without using the URL configured for the transport, so
this default doesn't matter when it's set. In other cases, this
will've needed to've been set explicitly before, so changing the
default should be quite benign.

Also I added a comment earlier to a Dockerfile but didn't get it into
somewhere, so tossing that in this PR.