feat: manage_programs now creates organizations if need be
This command is part of the build process for Master's Integration
Environments, the build process of which is described here:
https://openedx.atlassian.net/wiki/spaces/PT/pages/1028489324/Master+s+Integration+Environment
The command did not previously create organizations but instead failed
when no authoring org.s were previously put in place by another
command. This is because Integration Environments writ large only
support a single authoring organization; we only authorize a user to
use one at a time. However, to have better advance visibility when we
wouldn't be able to create an Integration Environment in response to a
partner request, we'd like to add a wider variety of programs to our
auto-built canary environment.