fix: broken setUp method due to freezegun upgrade
The version upgrade from 1.2.1 to 1.2.2 of freezegun breaks tests in the ReportsListViewTest class. This commits fixes the issue.

Version 1.2.2 of freezegun includes a change to "pass through all setUp and tearDown arguments". See the CHANGELOG here: https://github.com/spulec/freezegun/blob/master/CHANGELOG and the associated commit here: https://github.com/spulec/freezegun/commit/33fdce860fc72d7cea1c61c6e806245cc0dc96dc.

The ReportsListViewTest defined a tearDown method that was improperly defined. It looked like a mixture of a tearDown method and a tearDownClass method. It was named tearDown, but it was decorated with the @classmethod decorator and accepted cls as a parameter, which is how a tearDownClass method is written. However, it was named tearDown.

Changing the tearDown method to a tearDownClass method causes numerous errors in the test suite due to failed uniqueness constraints. Therefore, the tearDown method was implemented as a tearDown. This fixes the test failures.