feat: enable course role management for Program Console
This commit enables course role management via CSV for use by the Program Console frontend application. Previously, this feature was only enabled for the Registrar REST API, which is used by program partners to manage course and program enrollments.

Now, provided that the appropriate user group is added to the enable_course_role_management waffle flag, any user that is a member of said user group can use the Program Console to download course enrollments and see whether enrolled learners in the course are course staff and to upload course enrollments to effect a change in a learner's course staff status.